Reorganize drawing into Passes (#10)
Previously material binding, pipeline state objects, and root signatures were deeply coupled. This would have been OK for an entity focused engine, but makes no sense when we're primarily working in discrete passes with not many entities at once.

I've decoupled Root Sig/PSO from the material structure. Now "Pass" and its derived class "GraphicsPass" handle queries for parameter indices and binding shaders to the pipeline.

This was done to be extensible to the incoming "ComputePass" which will be done in the same way.

* Compute Pass Creation, UAV Parsing

* ComputePass -> Pass

Let's just do it right the first time.

* Create initial shader hierarchy

* Fix abstract base class issue.

* Create Pass class hierarchy

* Replace bloated material class with Pass structure

* Replace bloated MaterialType class with more lightweight Material

Going by Pass will help us decouple the pipeline from entity drawing, which is not the primary focus of this engine.